tests/wpt/mozilla/tests/mozilla/mozbrowser/private_browsing.html

<head>
  <meta charset="utf8" />
  <title>Private browsing</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    async_test(function(t) {
      var privateFrame = document.createElement("iframe");
      privateFrame.mozbrowser = true;
      privateFrame.mozprivatebrowsing = true;

      var gotGrandchildResult = false;
      privateFrame.addEventListener("mozbrowsershowmodalprompt", t.step_func(e => {
        assert_equals(e.detail.message, 'private=active private=active');
        gotGrandchildResult = true;
      }));

      privateFrame.onload = t.step_func(function() {
        assert_true(gotGrandchildResult);

        var parent = privateFrame.parentNode;
        parent.removeChild(privateFrame);

        var iframe = document.createElement("iframe");
        var promptDisplay = false;
        iframe.mozbrowser = true;
        iframe.onload = t.step_func(function() {
          assert_true(promptDisplay);
          t.done();
        });
        iframe.addEventListener("mozbrowsershowmodalprompt", t.step_func(e => {
          promptDisplay = true;
          assert_equals(e.detail.message, "");
        }));
        iframe.src = "iframe_contentDocument_inner.html";
        parent.appendChild(iframe);
      });
      privateFrame.src = "iframe_privateContent_inner.html";
      document.body.appendChild(privateFrame);
    });
  </script>
</body>
